// src/payload-generated-schema.ts
export const rss = pgTable(
  'rss',
  {
    id: serial('id').primaryKey(),
    link: varchar('link').notNull(),
    data: jsonb('data'),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
  },
  (columns) => ({
    rss_updated_at_idx: index('rss_updated_at_idx').on(columns.updatedAt),
    rss_created_at_idx: index('rss_created_at_idx').on(columns.createdAt),
  }),
)

export const rss_data = pgTable(
  'rss_data',
  {
    id: serial('id').primaryKey(),
    rss: integer('rss_id').references(() => rss.id, {
      onDelete: 'set null',
    }),
    data: jsonb('data'),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
  },
  (columns) => ({
    rss_data_rss_idx: index('rss_data_rss_idx').on(columns.rss),
    rss_data_updated_at_idx: index('rss_data_updated_at_idx').on(columns.updatedAt),
    rss_data_created_at_idx: index('rss_data_created_at_idx').on(columns.createdAt),
  }),
)

export const rss_translate = pgTable(
  'rss_translate',
  {
    id: serial('id').primaryKey(),
    data: jsonb('data'),
    rssData: integer('rss_data_id').references(() => rss_data.id, {
      onDelete: 'set null',
    }),
    updatedAt: timestamp('updated_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
    createdAt: timestamp('created_at', { mode: 'string', withTimezone: true, precision: 3 })
      .defaultNow()
      .notNull(),
  },
  (columns) => ({
    rss_translate_rss_data_idx: index('rss_translate_rss_data_idx').on(columns.rssData),
    rss_translate_updated_at_idx: index('rss_translate_updated_at_idx').on(columns.updatedAt),
    rss_translate_created_at_idx: index('rss_translate_created_at_idx').on(columns.createdAt),
  }),
)